Move authMe to the /api/user handler (#93)
