+++ backend/internal/handler/auth_test.go
@@ -10,7 +10,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/zouipo/yumsday/backend/internal/constant"
 	"github.com/zouipo/yumsday/backend/internal/ctx"
 	"github.com/zouipo/yumsday/backend/internal/dto"
 	customErrors "github.com/zouipo/yumsday/backend/internal/error"
@@ -20,7 +19,6 @@ import (
 
 var (
 	loginRoute  = "/auth/login"
-	authMeRoute = "/auth/me"
 	logoutRoute = "/auth/logout"
 
 	username      = "username"
@@ -213,112 +211,6 @@ func TestPostLogin_GenericError(t *testing.T) {
 
 	if !strings.Contains(w.Body.String(), "an error occurred while checking credentials") {
 		t.Errorf("expected error message containing %q instead of %q", "an error occurred while checking credentials", w.Body.String())
-	}
-}
-
-/*** TESTS AuthMe ***/
-
-func TestAuthMe_Success(t *testing.T) {
-	avatar := enum.Avatar1
-	authenticatedUser := &model.User{
-		ID:        42,
-		Username:  username,
-		AppAdmin:  true,
-		CreatedAt: time.Now().UTC(),
-		Avatar:    &avatar,
-		Language:  enum.English,
-		AppTheme:  enum.Light,
-	}
-	mockService := &mockAuthService{}
-	handler := NewAuthHandler(mockService)
-
-	r := httptest.NewRequest(http.MethodGet, authMeRoute, nil)
-	r = r.WithContext(context.WithValue(r.Context(), ctx.UserCtxKey{}, authenticatedUser))
-	w := httptest.NewRecorder()
-
-	handler.authMe(w, r)
-
-	if w.Code != http.StatusOK {
-		t.Errorf("expected status %d instead of %d", http.StatusOK, w.Code)
-	}
-
-	if w.Header().Get(constant.CONTENT_TYPE_HEADER) != constant.CONTENT_TYPE_VALUE {
-		t.Errorf(
-			"expected header %q to be %q instead of %q",
-			constant.CONTENT_TYPE_HEADER,
-			constant.CONTENT_TYPE_VALUE,
-			w.Header().Get(constant.CONTENT_TYPE_HEADER),
-		)
-	}
-
-	var userDto dto.UserDto
-	if err := json.Unmarshal(w.Body.Bytes(), &userDto); err != nil {
-		t.Fatalf("expected valid user JSON response, got error: %v", err)
-	}
-
-	if err := compareUserToUserDto(&userDto, authenticatedUser); err != nil {
-		t.Errorf("response body mismatch: %v", err)
-	}
-
-	if mockService.authCalls != 0 {
-		t.Errorf("expected auth calls 0 instead of %d", mockService.authCalls)
-	}
-
-	if mockService.logoutCalls != 0 {
-		t.Errorf("expected logout calls 0 instead of %d", mockService.logoutCalls)
-	}
-}
-
-func TestAuthMe_MissingUserInContext(t *testing.T) {
-	mockService := &mockAuthService{}
-	handler := NewAuthHandler(mockService)
-
-	r := httptest.NewRequest(http.MethodGet, authMeRoute, nil)
-	w := httptest.NewRecorder()
-
-	handler.authMe(w, r)
-
-	if w.Code != http.StatusInternalServerError {
-		t.Errorf("expected status %d instead of %d", http.StatusInternalServerError, w.Code)
-	}
-
-	if !strings.Contains(w.Body.String(), http.StatusText(http.StatusInternalServerError)) {
-		t.Errorf("expected error message containing %q instead of %q", http.StatusText(http.StatusInternalServerError), w.Body.String())
-	}
-
-	if mockService.authCalls != 0 {
-		t.Errorf("expected auth calls 0 instead of %d", mockService.authCalls)
-	}
-
-	if mockService.logoutCalls != 0 {
-		t.Errorf("expected logout calls 0 instead of %d", mockService.logoutCalls)
-	}
-}
-
-func TestAuthMe_InvalidUserTypeInContext(t *testing.T) {
-	mockService := &mockAuthService{}
-	handler := NewAuthHandler(mockService)
-
-	r := httptest.NewRequest(http.MethodGet, authMeRoute, nil)
-	r = r.WithContext(context.WithValue(r.Context(), ctx.UserCtxKey{}, "not-a-user"))
-	w := httptest.NewRecorder()
-
-	handler.authMe(w, r)
-
-	if w.Code != http.StatusInternalServerError {
-		t.Errorf("expected status %d instead of %d", http.StatusInternalServerError, w.Code)
-	}
-
-	if !strings.Contains(w.Body.String(), http.StatusText(http.StatusInternalServerError)) {
-		t.Errorf("expected error message containing %q instead of %q", http.StatusText(http.StatusInternalServerError), w.Body.String())
-	}
-
-	if mockService.authCalls != 0 {
-		t.Errorf("expected auth calls 0 instead of %d", mockService.authCalls)
-	}
-
-	if mockService.logoutCalls != 0 {
-		t.Errorf("expected logout calls 0 instead of %d", mockService.logoutCalls)
 	}
 }
